Move Frame related types into Frame Component
