+++ src/components/Frame/components/ContextualSaveBar/index.ts
@@ -1,4 +1,3 @@
 import ContextualSaveBar from './ContextualSaveBar';
 
-export {Props} from './ContextualSaveBar';
 export default ContextualSaveBar;
