+++ src/components/Frame/components/Toast/Toast.tsx
@@ -5,14 +5,13 @@ import {Key} from '../../../../types';
 
 import Icon from '../../../Icon';
 import KeypressListener from '../../../KeypressListener';
-import {DEFAULT_TOAST_DURATION} from '../../../Toast';
-import {ToastProps} from '../../../types';
+import {ToastProps} from '../../types';
 
 import * as styles from './Toast.scss';
 
-export type Props = ToastProps;
+export const DEFAULT_TOAST_DURATION = 5000;
 
-export default class Toast extends React.Component<Props, never> {
+export default class Toast extends React.Component<ToastProps, never> {
   private timer?: number;
 
   componentDidUpdate() {
