+++ src/components/Frame/components/index.ts
@@ -1,10 +1,7 @@
-export {default as Toast, Props as ToastProps} from './Toast';
+export {default as Toast, DEFAULT_TOAST_DURATION} from './Toast';
 export {
   default as ToastManager,
   Props as ToastManagerProps,
 } from './ToastManager';
 export {default as Loading, Props as LoadingProps} from './Loading';
-export {
-  default as ContextualSaveBar,
-  Props as ContextualSaveBarProps,
-} from './ContextualSaveBar';
+export {default as ContextualSaveBar} from './ContextualSaveBar';
