+++ src/components/Frame/index.ts
@@ -1,4 +1,14 @@
 import Frame from './Frame';
 
 export {Props} from './Frame';
+
+export {DEFAULT_TOAST_DURATION} from './components';
+
+export {
+  ContextualSaveBarProps,
+  FrameContext,
+  frameContextTypes,
+  ToastProps,
+} from './types';
+
 export default Frame;
