+++ src/components/Toast/Toast.tsx
@@ -2,10 +2,13 @@ import * as React from 'react';
 import {createUniqueIDFactory} from '@shopify/javascript-utilities/other';
 import {Flash as AppBridgeToast} from '@shopify/app-bridge/actions';
 
-import {FrameContext, frameContextTypes, ToastProps} from '../types';
+import {
+  DEFAULT_TOAST_DURATION,
+  FrameContext,
+  frameContextTypes,
+  ToastProps,
+} from '../Frame';
 import {withAppProvider, WithAppProviderProps} from '../AppProvider';
-
-export const DEFAULT_TOAST_DURATION = 5000;
 
 const createId = createUniqueIDFactory('Toast');
