+++ src/components/Toast/index.ts
@@ -1,4 +1,3 @@
 import Toast from './Toast';
 
-export {DEFAULT_TOAST_DURATION} from './Toast';
 export default Toast;
