+++ src/components/index.ts
@@ -115,7 +115,11 @@ export {default as Form, Props as FormProps} from './Form';
 
 export {default as FormLayout, Props as FormLayoutProps} from './FormLayout';
 
-export {default as Frame, Props as FrameProps} from './Frame';
+export {
+  default as Frame,
+  Props as FrameProps,
+  DEFAULT_TOAST_DURATION,
+} from './Frame';
 
 export {default as Heading, Props as HeadingProps} from './Heading';
 
@@ -237,7 +241,7 @@ export {
   Context as ThemeProviderContext,
 } from './ThemeProvider';
 
-export {default as Toast, DEFAULT_TOAST_DURATION} from './Toast';
+export {default as Toast} from './Toast';
 
 export {default as Tooltip, Props as TooltipProps} from './Tooltip';
