+++ src/components/types.ts
@@ -1,5 +1,3 @@
-import * as PropTypes from 'prop-types';
-
 import {Context as AppProviderContext} from './AppProvider';
 import {Context as ThemeProviderContext} from './ThemeProvider';
 
@@ -7,70 +5,4 @@ export interface PolarisContext
   extends AppProviderContext,
     ThemeProviderContext {}
 
-export interface FrameManager {
-  showToast(toast: {id: string} & ToastProps): void;
-  hideToast(toast: {id: string}): void;
-  setContextualSaveBar(props: ContextualSaveBarProps): void;
-  removeContextualSaveBar(): void;
-  startLoading(): void;
-  stopLoading(): void;
-}
-
-export interface FrameContext {
-  frame: FrameManager;
-}
-
-export const frameContextTypes = {
-  frame: PropTypes.object,
-};
-
 export type TransitionStatus = 'entering' | 'entered' | 'exiting' | 'exited';
-
-// FROM ContextualSaveBar
-
-export interface Action {
-  /** A destination to link to */
-  url?: string;
-  /** Content the action displays */
-  content?: string;
-  /** Should a spinner be displayed */
-  loading?: boolean;
-  /** Should the action be disabled */
-  disabled?: boolean;
-  /** Callback when an action takes place */
-  onAction?(): void;
-}
-
-interface DiscardActionProps {
-  /** Whether to show a modal confirming the discard action */
-  discardConfirmationModal?: boolean;
-}
-
-type CombinedActionProps = DiscardActionProps & Action;
-
-export interface ContextualSaveBarProps {
-  /** Extend the contents section to be flush with the left edge  */
-  alignContentFlush?: boolean;
-  /** Accepts a string of content that will be rendered to the left of the actions */
-  message?: string;
-  /** Save or commit contextual save bar action with text defaulting to 'Save' */
-  saveAction?: Action;
-  /** Discard or cancel contextual save bar action with text defaulting to 'Discard' */
-  discardAction?: CombinedActionProps;
-}
-
-// Toast
-
-export interface ToastProps {
-  /** The content that should appear in the toast message */
-  content: string;
-  /**
-   * The length of time in milliseconds the toast message should persist
-   * @default 5000
-   */
-  duration?: number;
-  /** Display an error toast. */
-  error?: boolean;
-  /** Callback when the dismiss icon is clicked */
-  onDismiss(): void;
-}
